Split EventStoreDB Card Domain into slices
