+++ Sample/ECommerce/Carts/Carts/Carts.csproj
@@ -18,10 +18,10 @@
 
     <ItemGroup>
         <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-            <_Parameter1>$(MSBuildProjectName).Tests</_Parameter1>
+            <_Parameter1>$(AssemblyName).Tests</_Parameter1>
         </AssemblyAttribute>
         <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-            <_Parameter1>$(MSBuildProjectName).Api.Tests</_Parameter1>
+            <_Parameter1>$(AssemblyName).Api.Tests</_Parameter1>
         </AssemblyAttribute>
     </ItemGroup>
