+++ Sample/EventStoreDB/ECommerce/Carts/Carts.Api/Controllers/CartsController.cs
@@ -2,9 +2,12 @@ using System;
 using System.Threading.Tasks;
 using Ardalis.GuardClauses;
 using Carts.Api.Requests.Carts;
-using Carts.Carts.Projections;
+using Carts.Carts.GettingCartAtVersion;
+using Carts.Carts.GettingCartById;
+using Carts.Carts.GettingCartHistory;
+using Carts.Carts.GettingCarts;
+using Carts.Carts.Products;
 using Carts.Carts.Queries;
-using Carts.Carts.ValueObjects;
 using Microsoft.AspNetCore.Mvc;
 using Core.Commands;
 using Core.Ids;
@@ -12,7 +15,6 @@ using Core.Marten.Responses;
 using Core.Queries;
 using Core.Responses;
 using Marten.Pagination;
-using Commands = Carts.Carts.Commands;
 
 namespace Carts.Api.Controllers
 {
@@ -38,13 +40,13 @@ namespace Carts.Api.Controllers
         }
 
         [HttpPost]
-        public async Task<IActionResult> InitCart([FromBody] InitCartRequest request)
+        public async Task<IActionResult> InitializeCart([FromBody] InitializeCartRequest request)
         {
             Guard.Against.Null(request, nameof(request));
 
             var cartId = idGenerator.New();
 
-            var command = Commands.InitCart.Create(
+            var command = Carts.InitializingCart.InitializeCart.Create(
                 cartId,
                 request.ClientId
             );
@@ -60,7 +62,7 @@ namespace Carts.Api.Controllers
             Guard.Against.Null(request, nameof(request));
             Guard.Against.Null(request.ProductItem, nameof(request));
 
-            var command = Commands.AddProduct.Create(
+            var command = Carts.AddingProduct.AddProduct.Create(
                 id,
                 ProductItem.Create(
                     request.ProductItem.ProductId,
@@ -79,7 +81,7 @@ namespace Carts.Api.Controllers
             Guard.Against.Null(request, nameof(request));
             Guard.Against.Null(request.ProductItem, nameof(request));
 
-            var command = Commands.RemoveProduct.Create(
+            var command = Carts.RemovingProduct.RemoveProduct.Create(
                 id,
                 PricedProductItem.Create(
                     request.ProductItem.ProductId,
@@ -96,7 +98,7 @@ namespace Carts.Api.Controllers
         [HttpPut("{id}/confirmation")]
         public async Task<IActionResult> ConfirmCart(Guid id)
         {
-            var command = Commands.ConfirmCart.Create(
+            var command = Carts.ConfirmingCart.ConfirmCart.Create(
                 id
             );
 
@@ -127,12 +129,12 @@ namespace Carts.Api.Controllers
 
             return pagedList.ToResponse();
         }
-        //
-        // [HttpGet("{id}/versions")]
-        // public Task<CartDetails> GetVersion(Guid id, [FromQuery] GetCartAtVersion query)
-        // {
-        //     Guard.Against.Null(query, nameof(query));
-        //     return queryBus.Send<GetCartAtVersion, CartDetails>(GetCartAtVersion.Create(id, query.Version));
-        // }
+
+        [HttpGet("{id}/versions")]
+        public Task<CartDetails> GetVersion(Guid id, [FromQuery] GetCartAtVersion query)
+        {
+            Guard.Against.Null(query, nameof(query));
+            return queryBus.Send<GetCartAtVersion, CartDetails>(GetCartAtVersion.Create(id, query.Version));
+        }
     }
 }
