--- Sample/EventStoreDB/ECommerce/Carts/Carts.Api/Requests/Carts/InitCartRequest.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace Carts.Api.Requests.Carts
-{
-    public class InitCartRequest
-    {
-        public Guid ClientId { get; set; }
-    }
-}
