+++ Sample/EventStoreDB/ECommerce/Carts/Carts.Api/Requests/Carts/InitializeCartRequest.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Carts.Api.Requests.Carts
+{
+    public class InitializeCartRequest
+    {
+        public Guid ClientId { get; set; }
+    }
+}
