+++ Sample/EventStoreDB/ECommerce/Carts/Carts.Tests/Carts/InitializingCart/InitializeCartTests.cs
@@ -1,0 +1,36 @@
+using System;
+using Carts.Carts;
+using Carts.Tests.Extensions.Reservations;
+using Xunit;
+
+namespace Carts.Tests.Carts.InitializingCart
+{
+    public class InitializeCartTests
+    {
+        [Fact]
+        public void ForValidParams_ShouldCreateCartWithPendingStatus()
+        {
+            // Given
+            var cartId = Guid.NewGuid();
+            var clientId = Guid.NewGuid();
+
+            // When
+            var cart = Cart.Initialize(
+                cartId,
+                clientId
+            );
+
+            // Then
+
+            cart
+                .IsInitializedCartWith(
+                    cartId,
+                    clientId
+                )
+                .HasCartInitializedEventWith(
+                    cartId,
+                    clientId
+                );
+        }
+    }
+}
