+++ Sample/EventStoreDB/ECommerce/Carts/Carts/Carts.csproj
@@ -7,23 +7,23 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <ProjectReference Include="..\..\..\..\..\Core\Core.csproj"/>
-        <ProjectReference Include="..\..\..\..\..\Core.EventStoreDB\Core.EventStoreDB.csproj"/>
+        <ProjectReference Include="..\..\..\..\..\Core\Core.csproj" />
+        <ProjectReference Include="..\..\..\..\..\Core.EventStoreDB\Core.EventStoreDB.csproj" />
         <!-- Document part used for projections -->
-        <ProjectReference Include="..\..\..\..\..\Core.Marten\Core.Marten.csproj"/>
+        <ProjectReference Include="..\..\..\..\..\Core.Marten\Core.Marten.csproj" />
     </ItemGroup>
 
     <ItemGroup>
-        <PackageReference Include="Ardalis.GuardClauses" Version="3.2.0"/>
-        <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="5.0.0"/>
+        <PackageReference Include="Ardalis.GuardClauses" Version="3.2.0" />
+        <PackageReference Include="Microsoft.Extensions.Configuration.Abstractions" Version="5.0.0" />
     </ItemGroup>
 
     <ItemGroup>
         <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-            <_Parameter1>$(MSBuildProjectName).Tests</_Parameter1>
+            <_Parameter1>$(AssemblyName).Tests</_Parameter1>
         </AssemblyAttribute>
         <AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleTo">
-            <_Parameter1>$(MSBuildProjectName).Api.Tests</_Parameter1>
+            <_Parameter1>$(AssemblyName).Api.Tests</_Parameter1>
         </AssemblyAttribute>
     </ItemGroup>
