+++ Sample/EventStoreDB/ECommerce/Carts/Carts/Carts/AddingProduct/ProductAdded.cs
@@ -1,0 +1,28 @@
+using System;
+using Ardalis.GuardClauses;
+using Carts.Carts.Products;
+using Core.Events;
+
+namespace Carts.Carts.AddingProduct
+{
+    public class ProductAdded: IEvent
+    {
+        public Guid CartId { get; }
+
+        public PricedProductItem ProductItem { get; }
+
+        private ProductAdded(Guid cartId, PricedProductItem productItem)
+        {
+            CartId = cartId;
+            ProductItem = productItem;
+        }
+
+        public static ProductAdded Create(Guid cartId, PricedProductItem productItem)
+        {
+            Guard.Against.Default(cartId, nameof(cartId));
+            Guard.Against.Null(productItem, nameof(productItem));
+
+            return new ProductAdded(cartId, productItem);
+        }
+    }
+}
