+++ Sample/EventStoreDB/ECommerce/Carts/Carts/Carts/Cart.cs
@@ -2,8 +2,11 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using Ardalis.GuardClauses;
-using Carts.Carts.Events;
-using Carts.Carts.ValueObjects;
+using Carts.Carts.AddingProduct;
+using Carts.Carts.ConfirmingCart;
+using Carts.Carts.InitializingCart;
+using Carts.Carts.Products;
+using Carts.Carts.RemovingProduct;
 using Carts.Pricing;
 using Core.Aggregates;
 using Core.Extensions;
