--- Sample/EventStoreDB/ECommerce/Carts/Carts/Carts/Commands/RemoveProduct.cs
@@ -1,24 +1,0 @@
-using System;
-using Carts.Carts.ValueObjects;
-using Core.Commands;
-
-namespace Carts.Carts.Commands
-{
-    public class RemoveProduct: ICommand
-    {
-        public Guid CartId { get; }
-
-        public PricedProductItem ProductItem { get; }
-
-        private RemoveProduct(Guid cardId, PricedProductItem productItem)
-        {
-            CartId = cardId;
-            ProductItem = productItem;
-        }
-
-        public static RemoveProduct Create(Guid cardId, PricedProductItem productItem)
-        {
-            return new RemoveProduct(cardId, productItem);
-        }
-    }
-}
