+++ Sample/EventStoreDB/ECommerce/Carts/Carts/Carts/ConfirmingCart/CartConfirmed.cs
@@ -1,0 +1,20 @@
+using System;
+using Ardalis.GuardClauses;
+using Core.Events;
+
+namespace Carts.Carts.ConfirmingCart
+{
+    public record CartConfirmed(
+        Guid CartId,
+        DateTime ConfirmedAt
+    ): IEvent
+    {
+        public static CartConfirmed Create(Guid cartId, DateTime confirmedAt)
+        {
+            Guard.Against.Default(cartId, nameof(cartId));
+            Guard.Against.Default(confirmedAt, nameof(confirmedAt));
+
+            return new CartConfirmed(cartId, confirmedAt);
+        }
+    }
+}
