--- Sample/EventStoreDB/ECommerce/Carts/Carts/Carts/Events/CartConfirmed.cs
@@ -1,20 +1,0 @@
-using System;
-using Ardalis.GuardClauses;
-using Core.Events;
-
-namespace Carts.Carts.Events
-{
-    public record CartConfirmed(
-        Guid CartId,
-        DateTime ConfirmedAt
-    ): IEvent
-    {
-        public static CartConfirmed Create(Guid cartId, DateTime confirmedAt)
-        {
-            Guard.Against.Default(cartId, nameof(cartId));
-            Guard.Against.Default(confirmedAt, nameof(confirmedAt));
-
-            return new CartConfirmed(cartId, confirmedAt);
-        }
-    }
-}
